src/requirements.txt
requests
PyMuPDF
python-docx
python-dotenv